Refactored headless handling and support...
  - Added support for headless for both local
    and remote browsers
  - Moved headless specification from `BROWSER`
    to `HEADLESS`
  - Refactored to pass-through method of
    browser support creation
  - Removed support for non-Selenium
    Capybar browsers (drivers)
    (i.e. PhantomJS)
  - Moved browser creation to own support
    file
  - Updated README
