+++ Gemfile
@@ -5,8 +5,6 @@ source 'https://rubygems.org'
 gem 'bundler-audit', require: false
 gem 'capybara'
 gem 'parallel_tests'
-gem 'phantomjs-helper'
-gem 'poltergeist'
 gem 'rake'
 gem 'rspec'
 gem 'rspec-retry'
