+++ docker-compose.selenium.yml
@@ -3,6 +3,7 @@ services:
   browsertests:
     environment:
       - BROWSER=${BROWSER:-chrome}
+      - HEADLESS=${HEADLESS-false}
       - REMOTE=http://${SELENIUM_HOSTNAME:-seleniumbrowser}:4444/wd/hub
       - REMOTE_STATUS=http://${SELENIUM_HOSTNAME:-seleniumbrowser}:4444/wd/hub/status
     command: ./script/runtests
